Moved Calculator to project_1
