--- ProjectCalculator/src/Operation.java
@@ -1,5 +1,0 @@
-package ProjectCalculator.src;
-
-interface Operation {
-    double Calculate();
-}
