+++ project_1/src/Calculator.java
@@ -1,19 +1,22 @@
-package ProjectCalculator.src;
+package project_1.src;
 
 import java.util.HashMap;
 import java.util.Map;
 
 public class Calculator {
     private Map<Character, Operation> operationMap = new HashMap<>();
-    public Calculator() {
-        operationMap.put('+', new Addition());
-        operationMap.put('-', new Subtraction());
-        operationMap.put('*', new Multiplication());
-        operationMap.put('/', new Division());
+    public Calculator(double firstNum, double secondNum) {
+        operationMap.put('+', new Addition(firstNum, secondNum));
+        operationMap.put('-', new Subtraction(firstNum, secondNum));
+        operationMap.put('*', new Multiplication(firstNum, secondNum));
+        operationMap.put('/', new Division(firstNum, secondNum));
+        operationMap.put('%', new Modulo(firstNum, secondNum));
+        operationMap.put('^', new Exponential(firstNum, secondNum));
+        
     }
 
-    public double makeCalculation(double operand1, double operand2, char operation) {
+    public double Calculate(char operation) {
         Operation operationMapValue = operationMap.getOrDefault(operation,new NoOperationFound(operation));
-        return operationMapValue.calculateResult(operand1, operand2);
+        return operationMapValue.Calculate();
     }
 }
