+++ project_1/src/Division.java
@@ -1,9 +1,16 @@
-package ProjectCalculator.src;
+package project_1.src;
 
 public class Division implements Operation {
+    private final double firstNum;
+    private final double secondNum;
+    Division(double firstNum, double secondNum){
+        this.firstNum = firstNum;
+        this.secondNum = secondNum;
+    }
+        
     @Override
-    public double calculateResult(double left, double right) {
-        if (right == 0) throw new UnsupportedOperationException("Cannot divide by zero");
-        return left / right;
+    public double Calculate() {
+        if (secondNum == 0) throw new UnsupportedOperationException("Cannot divide by zero");
+        return firstNum / secondNum;
     }
 }
